Relations between entities and bookings validation Upgraded all annotations to the new PHP 8 attributes for better compatibility.
Added all types of relations between various entities:
- OneToOne: Bookings & Cars (each car can have only one booking. if the car gets deleted, the booking will be removed as well)
- OneToMany: User & Bookings (one user can have multiple bookings, but each booking has only one owner)
- ManyToOne: Plugs & Stations (multiple plugs can be a part of a station, but each plug can be part of only one station)
- ManyToMany: Users & Cars (each user can have multiple cars and each car can have multiple owners)

When making a booking, the user can only select a car that they own and that is not part of another booking. Also, plug IDs are now validated to ensure the website doesn't crash when the user attempts to make a reservation to an inexistent plug. This will be replaced however when the reservations will be made using the map and stations selector (TODO).